requirements.txt
Django==1.5.4
django-tastypie==0.10.0
mimeparse==0.1.3
